Move query string input and query bar top row to NP
