+++ src/legacy/core_plugins/data/public/index.scss
@@ -1,5 +1,3 @@
 @import 'src/legacy/ui/public/styles/styling_constants';
 
-@import './query/query_bar/index';
-
 @import '../../../../plugins/data/public/ui/index'
