--- src/legacy/core_plugins/data/public/query/query_bar/_index.scss
@@ -1,4 +1,0 @@
-// SASSTODO: Formalize this color in Kibana's styling constants
-$typeaheadConjunctionColor: #7800A6;
-
-@import './components/index';
