+++ src/plugins/data/public/index_patterns/lib/index.ts
@@ -21,3 +21,4 @@ export { getTitle } from './get_title';
 export * from './types';
 export { validateIndexPattern } from './validate_index_pattern';
 export { IndexPatternMissingIndices } from './errors';
+export { getFromSavedObject } from './get_from_saved_object';
