+++ src/plugins/data/public/ui/_index.scss
@@ -4,3 +4,5 @@
 @import './typeahead/index';
 
 @import './saved_query_management/index';
+
+@import './query_string_input/index';
