+++ src/plugins/data/public/ui/index.ts
@@ -21,7 +21,9 @@ export { SuggestionsComponent } from './typeahead/suggestions_component';
 export { IndexPatternSelect } from './index_pattern_select';
 export { FilterBar } from './filter_bar';
 export { applyFiltersPopover } from './apply_filters';
+export { QueryStringInput } from './query_string_input/query_string_input';
+
 // temp export - will be removed as final components are migrated to NP
-export { QueryLanguageSwitcher } from './query_string_input/language_switcher';
+export { QueryBarTopRow } from './query_string_input/query_bar_top_row';
 export { SavedQueryManagementComponent } from './saved_query_management';
 export { SaveQueryForm, SavedQueryMeta } from './saved_query_form';
