+++ src/plugins/data/public/ui/query_string_input/_query_bar.scss
