+++ src/plugins/data/public/ui/typeahead/_index.scss
@@ -1,1 +1,3 @@
+$typeaheadConjunctionColor: #7800A6;
+
 @import 'suggestion';
